貸出機能作成

+++ LibraryManagement/src/main/java/app/controller/MemberController.java
@@ -53,7 +53,7 @@ public class MemberController {
 
     // 会員詳細表示・編集画面
     @GetMapping("/edit/{id}")
-    public String showEditForm(@PathVariable Long id, Model model) {
+    public String showEditForm(@PathVariable Integer id, Model model) {
         Member member = memberService.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid member Id:" + id));
         model.addAttribute("member", member);
         return "member/edit";
@@ -61,7 +61,7 @@ public class MemberController {
 
     // 会員更新処理
     @PostMapping("/edit/{id}")
-    public String updateMember(@PathVariable Long id, @ModelAttribute Member member) {
+    public String updateMember(@PathVariable Integer id, @ModelAttribute Member member) {
         member.setMemberId(id);
         memberService.save(member);
         return "redirect:/member";
@@ -69,14 +69,14 @@ public class MemberController {
 
     // 会員削除処理
     @GetMapping("/delete/{id}")
-    public String deleteMember(@PathVariable Long id) {
+    public String deleteMember(@PathVariable Integer id) {
         memberService.deleteById(id);
         return "redirect:/member";
     }
 
  // 予約一覧表示
     @GetMapping("/member_reservation_list/{memberId}")
-    public String showReservationList(@PathVariable Long memberId, Model model) {
+    public String showReservationList(@PathVariable Integer memberId, Model model) {
         List<Reservation> reservations = reservationService.findByMemberId(memberId);
         Member member = memberService.findById(memberId).orElse(null);
         model.addAttribute("reservations", reservations);
